add non cgo build
